Reafactor ServiceModel (#19)
* Refactor ServiceModel class to improve performance and clarity

Simplified and optimized the ServiceModel constructor by leveraging caching, enhancing conditional checks, and restructuring property assignments. Overall, this refactoring cuts down on complexity and improves code readability, as well as potentially providing performance enhancements due to the added caching and streamlined conditional checks.

* Add performance attribute in README

An attribute emphasizing efficiency and speed of the service has been added in the README file. This update helps users understand that the service is designed with performance and speed in mind.

---------

Co-authored-by: david_smith <david_smith@sweetwater.com>