Add test to handle enum in EnumTest
A new test has been added to EnumTest.php to handle enum usage. The test checks if the name of UsesEnum created by make function is an instance of MockEnumString and is equal to its test value. This helps in validating the creation and assignment of enums through make method.