Bugfix Pass Enums Directly (#20)
* Improve enum check in ServiceModel

Updated the enum checking condition in ServiceModel class to include integer and string values. This enhancement allows enums to be passed directly without requiring extra conversions, improving overall code efficiency.

* Add test to handle enum in EnumTest

A new test has been added to EnumTest.php to handle enum usage. The test checks if the name of UsesEnum created by make function is an instance of MockEnumString and is equal to its test value. This helps in validating the creation and assignment of enums through make method.

---------

Co-authored-by: david_smith <david_smith@sweetwater.com>