Implement Caching (#23)
* Optimize reflection handling using caching

The ServiceModel has been optimized for efficiency by implementing a caching system. The ReflectionClass instances and properties are now cached to reduce the overhead of repeatedly generating instances that are identical, improving the overall performance of the application.

* Add Cache class to ServiceModel namespace

A new Cache class has been added to the Zerotoprod\ServiceModel namespace. The class includes methods for setting and getting key-value pairs in a cache array, and also includes a method for retrieving single instance of the Cache class, implementing Singleton Design Pattern.

* Update README to detail caching mechanism

The README file has been updated to include a detailed section on the caching mechanism used in the ServiceModel trait. It highlights its usage and implementation via a Singleton pattern, and how it improves performance by storing ReflectionClass and ReflectionProperty instances.

---------

Co-authored-by: david_smith <david_smith@sweetwater.com>