Revamp "Getting Started" section in README
This update significantly expands the "Getting Started" section of the README. It now includes detailed steps on installing the package with composer, using the trait in a model, and adding properties to a model that match the keys of your data. It also demonstrates how to use the `make()` and `factory()` methods.