Bugfix/return when items are not array or object (#25)
* Fix condition in ServiceModel constructor

Updated the condition check in the constructor of the ServiceModel class. This change ensures that the provided argument is an array or an object, further strengthening the validation check and preventing potential errors down the line..

* Add CustomArrayCast test class

A new test class, CustomArrayCast, has been added under the Casters directory. This class tests if an array can be cast into a specific object, specifically the ChildWithoutTrait, ensuring the casting functionality is working properly and improving the robustness of our code.

* Add more unit tests to ConstructorTest

Enhanced the ConstructorTest class by adding unit tests for scenarios where an object and a string are passed, ensuring that in every case it returns an instance of the NativeTypes class. These additional tests further verity the functionality and integrity of the code.

---------

Co-authored-by: david_smith <david_smith@sweetwater.com>