Chore/documentation (#28)
* Update README.md with factory support detail

The README.md description of the DTOs now highlights the support for factories. Previously, this information wasn't explicitly mentioned, potentially leaving out a significant feature of the package.

* Update description in composer.json with factory support detail

The description in composer.json about the DTOs now includes the support for factories. This was added to provide a more accurate description of the package functionalities, which previously hadn't explicitly highlighted this feature.

---------

Co-authored-by: david_smith <david_smith@sweetwater.com>