Add pickups attribute to Order model (#31)
The Order model in README.md has been updated to include a 'pickups' attribute. This change includes updates in multiple sections of the document - under the Order::make function, usage examples and in the Order class. The attribute addition aims to meet the new requirement of handling order pickup details.

Co-authored-by: david_smith <david_smith@sweetwater.com>