Feat/improve caching (#32)
* Implement 'remember' method in Cache class

A 'remember' method has been added to the Cache class. This method checks whether a cache exists for a particular key and generates a new one using a callable if it does not. This enhancement improves the caching mechanism and reduces unnecessary regeneration of cache for same keys.

* Refactor caching mechanism in ServiceModel

The caching mechanism in the ServiceModel class has been refactored to use the 'remember' method from the Cache class. This change eliminates repetitive cache checks and simplifies the code by automatically generating cache for a key if it doesn't exist. It also results in a significant reduction in verbose code structure, making it cleaner and more readable.

* Refactor enum_exists check in ServiceModel

The 'enum_exists' check within the ServiceModel class has been adjusted to include a null value guard. This reduces unnecessary Cache 'remember' calls, which improves performance by avoiding redundant checks and caching operations. An emphasis has been placed on a more efficient and cleaner code structure.

* Rename argument in Cache 'remember' method

The 'remember' method in the Cache class has replaced the argument name 'valueGenerator' with 'callable'. This provides a more accurate description for all sorts of functions (not just those generating values), thereby enhancing readability and providing clarity in the function's usage.

* Add PhpUndefinedFunctionInspection to ServiceModel.php

Inserted 'PhpUndefinedFunctionInspection' tag to suppress the warning of potentially undefined functions in ServiceModel.php. This change improves code readability and decreases potential distractions due to warning messages while working with the ServiceModel file.

* Update expected exception in EnumTest

Revised the expected exception in the EnumTest from 'Error' to 'Throwable'. The change ensures that our test correctly handles all forms of program interruptions, enhancing its effectiveness and reliability.

---------

Co-authored-by: david_smith <david_smith@sweetwater.com>