Add Infection for mutation testing in Github workflow
Infection, a mutation testing framework for PHP, has been added to Github workflow. This will operate alongside existing tests to assess the quality of test suites. With Infection, the software's ability to detect problems, even when parts of the code are altered randomly, is enhanced to deliver a more robust application.