Feat/infection (#36)
* Add Infection for mutation testing in Github workflow

Infection, a mutation testing framework for PHP, has been added to Github workflow. This will operate alongside existing tests to assess the quality of test suites. With Infection, the software's ability to detect problems, even when parts of the code are altered randomly, is enhanced to deliver a more robust application.

* Add Infection configuration file for mutation testing

Infection, a mutation testing framework for PHP, is now included in our Github workflow. Alongside our existing tests, it will help evaluate the effectiveness of our test suites. This involves testing how well our software can detect bugs when random alterations are made to the code, contributing to an overall improvement in the robustness of our application.

* Ignore Infection output file in .gitignore

The Infection output file, infection.html, has been added to .gitignore. This is because Infection, a mutation testing framework for PHP, has been integrated into our development workflow. Ignoring the output file helps maintain a clean project directory, especially as this file is continually regenerated during testing.

* Add EmptyClass test case in CastTest

Added a new test case in Unit/CastTest.php to check instances of the EmptyClass model. This ensures that the EmptyClass model functions as expected, particularly with respect to non-existent attributes (in this case, the 'test' attribute).

* Add EmptyClass and CacheTest units

Introduced a new EmptyClass to test ServiceModel. Also, added unit tests for the Cache module in CacheTest. This includes tests for getting and setting cache values and handling non-existent keys.

* Add "PublicVisibility" unit in infection.json5

Added a new testing unit, "PublicVisibility", in infection.json5. The unit assists in ignoring a definition in the Zerotoprod\\ServiceModel\\Factory, enabling more effective control over testing procedures. This builds on broadening the scope and capabilities of our testing framework.

---------

Co-authored-by: david_smith <david_smith@sweetwater.com>