Implement API key for Infection Dashboard (#37)
* Implement API key for Infection Dashboard

Added an environment variable to the PHP GitHub workflow for the Infection Dashboard API key. This implementation will enhance the authentication for the Infection dashboard, enabling secure report updates and enhancing the integration security.

* Update infection.json5 configuration

Added a "stryker" configuration to the "logs" section in infection.json5 file. This will direct any "stryker" badges to be linked with a regex statement, aiding in clearer filtering and sorting on the dashboard.

---------

Co-authored-by: david_smith <david_smith@sweetwater.com>