Feat/stryker integration (#38)
* Add json log to infection.json5 configuration file

A "json" log link is added to the "logs" section in the infection.json5 configuration file. This update will enable the 'json' log type and increase functionality, achieving better sorting and filtering on the managing dashboard.

* Update Infection workflow to send reports to dashboard

The current Infection workflow is updated to not only run Infection but also send a JSON report of the results to the configured dashboard. This enhances analysis capabilities of mutation results by providing sorting and filtering functionalities on the dashboard.

* Add mutation testing badge to README.

A mutation testing badge from Stryker Mutator has been added to the README file. This change provides a clear and quick indicator of the mutation testing status across the project, enhancing visibility and understanding of code quality.

* Update Stryker Mutator report URL in GitHub actions.

The Stryker Mutator report URL in the GitHub Actions workflow file "php.yml" was updated. Instead of using a variable for the URL, the URL to the Stryker Mutator report for the "zero-to-prod/main" project path is now hardcoded.

* Update Stryker Mutator report URL in GitHub actions

The workflow file 'php.yml' has been updated to change the Stryker Mutator report URL. The URL has been hardcoded to point to the 'zero-to-prod/service-model/main' project path on the Stryker Mutator dashboard. This removes the use of a variable for the URL.

* Updated log configuration in infection.json5

Changed the log configuration in the infection.json5 file. The "json" field has been renamed to "summaryJson", and a new field "github" set to true has been added. This is to properly generate the report in the required format for GitHub.

* Change Stryker badge configuration in infection.json5

The Stryker badge configuration in the infection.json5 file has been updated. The matcher regex previously set only to recognize releases is now modified to match the 'main' branch, allowing proper generation of badges for 'main' branch.

* Remove reporting step in Infection test workflow

The step in the GitHub Actions workflow for running Infection tests, which previously reported the results to Stryker dashboard, has been removed. Now the workflow just executes Infection tests with certain parameters, without reporting the results anywhere.

---------

Co-authored-by: david_smith <david_smith@sweetwater.com>