Update API key variable name in test workflow
In the GitHub Actions Workflow for running PHP Infection tests, the name of the environment variable storing the dashboard API key has been changed. It was originally INFECTION_DASHBOARD_API_KEY, but it has been renamed to STRYKER_DASHBOARD_API_KEY to match the service used for reporting test results.