Update API key variable name in PHP workflow (#41)
* Update API key variable name in PHP workflow

The environment variable for the API key in the GitHub Actions workflow was updated. It changed from INFECTION_DASHBOARD_API_KEY to STRYKER_DASHBOARD_API_KEY, reflecting the service used for reporting test results.

* Update API key usage in PHP workflow

The upgrade in the GitHub Actions workflow is to improve the call to the API key. Instead of the previous method, it now uses echo to print out the STRYKER_DASHBOARD_API_KEY which is consistent with our new testing service - Stryker.

* Simplify Stryker API key call in PHP workflow

The commit refines the API key call in the PHP workflow. The echo command for STRYKER_DASHBOARD_API_KEY was removed as this was unnecessary and potentially insecure. This increases the efficiency and security of the workflow in our move to the new testing service, Stryker.

---------

Co-authored-by: david_smith <david_smith@sweetwater.com>