Update README.md and rename methods for more expressivity
Changed the 'set' methods in 'ToCarbon' and 'CastToCollection' classes to 'parse', improving semantic expressivity. Removed a badge from README.md and added a detailed guide to upgrade to version 2.0.0. This guide consists of important changes including updated attribute namespaces, the replacement of 'CanCast' interface with 'CanParse', and changing 'set' methods to 'parse' in classes.