Refactor CanCast interface to Contracts namespace
The CanCast interface has been moved to the ServiceModel\Contracts namespace in this revision. This change affects several files where the CanCast interface is imported. As a result, the import statements have been updated in all the respective files - including tests and the README - to reflect this new structure. This reorganization enhances code clarity and project structure.