Feat/v2 (#42)
* Move casting related classes to Attributes namespace

In this commit, all the casting related classes such as Cast, CastToArray and CastToClasses have been moved to a new Attributes namespace. This reorganization helps to group related entities together, make the project structure more maintainable and improve the readability of the code. README and test cases have been updated to incorporate these changes.

* Refactor CanCast interface to Contracts namespace

The CanCast interface has been moved to the ServiceModel\Contracts namespace in this revision. This change affects several files where the CanCast interface is imported. As a result, the import statements have been updated in all the respective files - including tests and the README - to reflect this new structure. This reorganization enhances code clarity and project structure.

* has been refactored to instantiate the Cache class under the new namespace, Zerotoprod\ServiceModel\Cache. Second, the Cache class has been moved from src directory to a new Cache directory in the src directory for better organization of the project structure.

* Refactor object instantiation in ServiceModel

The constructor in ServiceModel has been refactored to use a static method for object instantiation. This update includes the introduction of the make function and the utilization of a Cache singleton. Also, the way of handling the $items variable was altered, improving code clarity and maintainability.

* Replace 'CanCast' interface with 'CanParse'

The 'CanCast' interface and all related 'set' methods have been replaced by 'CanParse' interface with 'parse' methods. This is a significant change to the structure of the codebase, aiming to improve code readability, and enhance semantic expressivity. Corresponding changes have been applied throughout the code in classes implementing the updated interface.

* Update README.md and rename methods for more expressivity

Changed the 'set' methods in 'ToCarbon' and 'CastToCollection' classes to 'parse', improving semantic expressivity. Removed a badge from README.md and added a detailed guide to upgrade to version 2.0.0. This guide consists of important changes including updated attribute namespaces, the replacement of 'CanCast' interface with 'CanParse', and changing 'set' methods to 'parse' in classes.

* Update .gitignore to include infection.json

Added "infection.json" to .gitignore. This change is to ensure that this system-generated JSON file, which is usually created as a part of running Infection - a PHP mutation testing framework, does not get accidentally committed to the repository.

* Remove environment variable from GitHub action

Removed the STRYKER_DASHBOARD_API_KEY environment variable from the GitHub action workflow. It appears that the dashboard API key is no longer necessary for running Infection, hence it has been removed for code simplification and maintenance.

* Simplify logs in infection.json5 configuration file

Removed 'summaryJson', 'github', and 'stryker' configuration from logs in infection.json5. This change makes the configuration simpler and easier to maintain, as these elements were identified as no longer necessary for running Infection.

* Add upgrade guide link to README

A link to the upgrade guide for transitioning from version 1.0.0 has been added to the 'Getting Started' section of the README. This was done to provide users with a simple way to access this essential information.

* Update README.md to clarify ServiceModel trait usage

The README instructions on how to create a model using the ServiceModel trait have been simplified and made clearer. This change specifically updates the phrasing to ensure it is understood that the model is the one using the trait.

---------

Co-authored-by: david_smith <david_smith@sweetwater.com>