CastMethod Attribute (#43)
* Introduce CastMethod and refactor variable usage

The import and usage of the CastMethod attribute has been added in the ServiceModel. Also, the variable `$cast_classname` has been renamed to a more generic `$attribute_argument_0` to improve clarity and maintainability. This variable now holds the first argument retrieved from reflection attributes, which improves flexibility in handling different cast methods.

* Add new CastMethod attribute in ServiceModel

A new attribute, CastMethod, has been added in the ServiceModel to manage different casting methods. Notably, the variable `$method_name` has been introduced for the attribute argument, enhancing clarity and maintainability of the code.

* Create new test files for CastMethod attribute

The CastMethod attribute has been created along with four new test files to validate its functionality. It manages different casting methods in the ServiceModel. The commit includes tests for handling specific casting like time to `TimeClass` instance and testing top-level value casting.

* Add Class Method Parsing section to README

A detailed explanation on how to use a class method for parsing values was added to README file. A particular focus was put on the use of the `CastMethod` attribute, which helps to leverage these methods, with addition of sample use case involving'method `set` in the `TimeClass`.

* Refactor method `set` in README for clarity

The class example method `set` in README was refactored to enhance readability. The instance variable was renamed from `$instance` to `$self` to match common PHP coding standards and to improve comprehension for code readers.

---------

Co-authored-by: david_smith <david_smith@sweetwater.com>