Add Codecov action to Github workflow (#44)
* Add Codecov action to Github workflow

This change adds a new step to the Github CI workflow to upload coverage reports to Codecov. It's essential for visualizing and tracking our testing coverage over time and ensures that we can easily check the test coverage of each pull request.

* Update PestPHP run command, relocate Codecov upload

PestPHP run command has been updated to generate a coverage report. In addition, the step that uploads coverage reports to Codecov has been moved right after the PestPHP execution, instead of after running Infection. This helps in providing timely and accurate coverage data for each conducted test.

* Update PestPHP command to generate coverage report

The command for running PestPHP tests has been updated to generate a coverage.xml file for more detailed test coverage report. Additionally, the code coverage reports are now uploaded to Codecov directly after the tests are run, improving the timeliness and accuracy of the reported data.

* Update PestPHP command in Github workflow
Updated the PestPHP command in the Github workflow to change the generation and location of the coverage.xml file. This is meant to provide a more detailed test coverage report. The change also streamlines the uploading of the code coverage report directly to Codecov, improving accuracy and expediency of the report generation.

* Improve test coverage reporting in Github workflow
The PestPHP command in the Github workflow file is updated to modify the generation and path of the coverage.xml file. The revision allows for a more detailed and punctual test coverage report by optimizing the upload process to Codecov.

* Update PestPHP command and upload path in Github workflow

The changes update the PestPHP command and the coverage report file path in the GitHub workflow file. This improvement helps generate a more detailed test coverage report and optimizes the upload process to Codecov by adjusting the location of the coverage.xml file.

* Fix syntax in PestPHP command in Github workflow

The PestPHP command syntax and the path to the coverage.xml file were corrected in the GitHub workflow. These corrections will ensure the successful generation of the detailed test coverage report and optimize its upload process to Codecov.

* Update PestPHP command and coverage file path in GitHub workflow

The PestPHP command syntax in the GitHub Actions workflow has been modified, along with the path to the coverage.xml file. This addresses a potential issue with coverage report generation and its transmission to the Codecov service.

* Update PestPHP command and coverage file path in GitHub workflow

The PestPHP command syntax in the GitHub Actions workflow has been modified, along with the path to the coverage.xml file. This addresses a potential issue with coverage report generation and its transmission to the Codecov service.

* Enable Xdebug coverage mode in GitHub workflow

Enabled Xdebug coverage mode in the PestPHP command run within the GitHub Actions workflow, to better handle generation of coverage reports. This amendment would also help in ensuring the correct file path while uploading the reports to Codecov.

* Remove file path specification for codecov

Removed file path specification from codecov action in GitHub workflows. This change enhances compatibility with how PestPHP command generates coverage reports and ensures correct file path when uploading these reports to codecov.

---------

Co-authored-by: david_smith <david_smith@sweetwater.com>