Feat/add type coverage (#48)
* Update PestPHP command in GitHub Workflow

PestPHP command within the GitHub workflow is updated to run tests in parallel and enforce 100% type coverage. This ensures a more strict and efficient testing regime, going forward.

* Add pest-plugin-type-coverage to composer.json

The pestphp/pest-plugin-type-coverage dependency has been added to the composer.json file. This plugin will help enforce type coverage in PestPHP tests, increasing code reliability and robustness.

* Update PHP base image in Dockerfile

The base image in the Dockerfile was changed from PHP 8.1 FPM to PHP 8.1 CLI. This is to improve the handling of command line interfaces within Docker applications.

* Add type ignore annotation to parse method

The pest-ignore-type annotation has been added to the parse method in the CanParse interface. This is to ignore type checking on the parameter to accommodate for more flexible function usage and implementation.

* Improve type hinting in Factory class methods

Updated the Factory methods to better use type hints and added return types to enhance coding robustness and readability. Specific changes include the make, getRawAttributes, and newModel methods where input types are now explicitly specified and a return type is present.

* Add return type for 'remember' method in Cache class

The 'remember' method in the Cache class has been updated to explicitly declare its return type. This enhances the robustness of the code and improves its readability by ensuring better type hinting and providing a more accurate understanding of the expected return type.

* Update PestPHP test and coverage steps in workflow

The PestPHP test and coverage steps have been separated out within the GitHub workflow. The XDEBUG_MODE environment variable is now explicitly set in each step. This allows for more detailed control and management of the testing and coverage environments.

---------

Co-authored-by: david_smith <david_smith@sweetwater.com>