Feat/add enum support for cast to classes (#49)
* Refactor `parse` method in `CastToClasses`

The `parse` method in `CastToClasses` was updated to improve type-checking and object instantiation. It now better handles different types of values, especially integers and strings, and resolves them using the `tryFrom` method if an enum exists. The change also handles array values more effectively.

* Correct and improve variable name in 'parse' method

The variable `$value` has been renamed to `$values` in 'parse' method of the `CastToArray` class to improve readability and clarity. The new name better reflects that this variable is intended to handle multiple items, being more descriptive and consistent with its usage in the method.

* Update Carbon time instances to custom instances

The commit has replaced the use of Carbon instances for time parsing with a custom class instance. A new 'created_at' time property has been added, marking an equivalent step to the earlier 'ordered_at' property. Some classes have undergone minor modifications in the way they parse values. This modification aims to boost code adaptability and consistency.

* Update parameter name in ToChild's parse method

The parameter name in the parse method of ToChild class has been updated from $value to $values to improve clarity. The same change has been reflected in the way this parameter is used inside the method. Updating the parameter name provides a clearer understanding that the method is intended to handle multiple values, enhancing code readability.

* Add MockStringEnum model for testing

A new model, MockStringEnum, is added under the test directory. This is primarily intended to be used for testing purposes. It defines a string enumerator with a single case 'test'.

* Improve clarity of test descriptions in CastToClassTest.php

Updated test descriptions in CastToClassTest.php for clearer understanding. Made descriptions more explicit to specify the action, intention and expected result of each test. Also added two more tests for casting to arrays of MockStringEnum for extensive coverage.

* "Reorder array check in CastToClasses

The check for array values in CastToClasses.php has been moved prior to other checks improving the code readability and ensuring appropriate casting. This prioritizes array type checking and thus enhances the overall efficiency of the casting process."

* Add new const variables and casting in CastToClass

Introduced new const variables 'enums' and 'enum_value' in CastToClass.php. Also added their respective casts using the 'CastToClasses' attribute, allowing instances of specific enumerated types to be cast directly into these new variables in a more efficient way.

* Rename parameter in CanParse interface

Changed parameter name from '$value' to '$values' in the parse method of the CanParse interface. This change is aimed to clarify that multiple values can be parsed in this method.

* Refactor parameter name in parse method

The parameter's name in the parse method of the CanParse interface has been refactored. The name has been changed from '$value' to '$values' to better indicate that the method supports parsing of multiple values.

* Add PHP inspection comments to casting classes

PHP inspection comments were introduced to suppress warnings from `PhpUndefinedClassInspection` and `PhpUndefinedMethodInspection` in `CastToClasses.php` and `CastToArray.php`. This ensures cleaner code without warnings about undefined classes or methods.

---------

Co-authored-by: david_smith <david_smith@sweetwater.com>