Chore/update factory tests (#52)
* Add additional tests to FactoryTest.php

Expand the scope of FactoryTest.php by including tests for expanding attributes and the default definition of a factory. These tests will help identify any errors that may arise when expanding the attributes of an object or when creating default factory objects.

* Simplify and streamline tests in PlainFactoryTest.php

The complexity of tests in PlainFactoryTest.php was unnecessary and has been greatly simplified. All previous setup, assertions and multiple checks have been replaced with a single, more efficient and straightforward test. This is easier to read and maintain, improving overall code quality.

* Add new test for definition override in PlainFactoryModel

This commit adds a new test in PlainFactoryModel to check for the 'definition override' function. It ensures that the 'setName' method can successfully update the 'name' attribute to a new value, in this case, 'test'. This additional test enhances the unit tests and ensures the function works properly.

* Update model reference in PlainFactoryModelFactory

The FactoryModel reference has been changed to PlainFactoryModel in PlainFactoryModelFactory. In addition, the definition method and the setName overloading method have been updated for 'name' attribute directly rather than using FactoryModel::name. This enhances readability by directly referring to attributes.

* Refactor PlainFactoryModel's constructor and attribute references

This update revises the constructor within PlainFactoryModel to accommodate attributes directly, removing the need for constant attribute references. The 'name' attribute has been handled directly in the constructor, improving code readability and maintainability.

---------

Co-authored-by: david_smith <david_smith@sweetwater.com>