Refactor argument name in CustomCaster class
The argument name in the constructor of CustomCaster class has been renamed from $value to $attribute_constructor_value. This change provides a clearer understanding of the role of the argument within the scope of the constructor. It also improves code readability by providing more specific context for the argument usage.