Feat/map output names (#60)
* Add `toResource` method to ServiceModel trait

Introduced the `toResource` function that creates an array representation of a service model. It leverages PHP's ReflectionClass to conditionally handle objects with and without attributes. This enhancement aids in better data handling and manipulation across the application.

* Add `Resource Support` section to README.md

A new `Resource Support` section has been added to the README, detailing the new ability to transform ServiceModels to associative arrays with snake_case keys, and the potential for user customization of this feature. This section provides code examples and explains how the `toResource()` method works, enhancing user understanding and ease of use.

* Add PhpUndefinedMethodInspection to ServiceModel

Added annotation to `toResource()` method in the ServiceModel to suppress the PhpUndefinedMethodInspection warning. This makes code more readable and prevents unnecessary warnings during code analysis.

* Add ToArray attribute to ServiceModel

Created a new `ToArray` attribute within the ServiceModel. This attribute defines a method that parses an array of values, handling instances of objects, and further parsing nested arrays to improve value extraction and processing in the ServiceModel.

* Add MapOutputNames attribute to ServiceModel

Introduced a new MapOutputNames attribute for the ServiceModel to map names. This attribute defines a constructor function that accepts a map. This will improve the direction of reading/writing data from/to different sources.

* Create MapOutputNamesClass for testing map output names

A new class, MapOutputNamesClass, is introduced to the testing directory for output names. This class utilizes the MapOutputNames attribute for handling different names in the ServiceModel. The addition of this attribute will facilitate improved data transfer in different sources.

* Add NestedOutputNamesClass for testing snake case mapping

NestedOutputNamesClass has been introduced in the test directory for verifying the functionality of output names' snake case mapping. By using MapOutputNames, the class tests how service model handles different case conversions, paving the way for more clear and consistent data transfer across different sections of the app.

* Add ToResource class with nested classes for testing

Created a new class named ToResource under the test models. This class includes NestedOutputNamesClass and other classes, which will be used to test how well ServiceModel can manage mappings between various case styles. This enhancement is crucial for streamlining data transfers within different sections of the application.

* Add ToResource unit tests

A new file, ToResourceTest.php, was created to introduce unit tests for the ToResource class. This includes tests for snake case naming conventions and nested class integration. These tests are important for verifying the correct mapping of data models in various case styles.

* Add ToSnakeCase attribute in ServiceModel

A new attribute, ToSnakeCase, has been added to Zerotoprod\ServiceModel\Attributes. This attribute allows for parsing of array values into snake case, a common naming convention in PHP. The addition of this attribute aids in maintaining consistent and readable code across the project.

* Update snake function type hinting in ToSnakeCase

Modified the snake function in the ToSnakeCase class to include type hinting. The parameters '$value' and 'delimiter' are now explicitly declared as strings, and the return type is explicitly declared as 'mixed'. This update will ensure better type safety and code readability.

* Commit message:
Ignore code coverage in MapOutputNames

Added annotations to ignore code coverage at the start and end of the MapOutputNames file. This will exclude code from code coverage metrics, and helps to focus the coverage report on code that actually needs testing.

* Fix code coverage annotation in MapOutputNames

Changed the incorrect code coverage annotation at the end of the MapOutputNames file from 'Start' to 'End'. This clarifies the range of ignored code coverage, ensuring accuracy in the coverage metrics.

---------

Co-authored-by: david_smith <david_smith@sweetwater.com>